+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * Refactored resource type to use an enum
 * Refactoring ideas into a data file and adding cancelled state
 * Refactoring build information out of epilogue component into utils module
+* Refactoring PointOfContact construction to utils module from contacts module
 
 ## [0.5.3] - 2023-12-30